Revise README with 'Money Heist' information
Updated README to include details about 'Money Heist'.

What I really appreciate in *Money Heist* is that every episode is very intense and unpredictable. The narrative is very thrilling due to the meticulous planning of **The Professor**, particularly when everything goes wrong. The characters are all different and have their individual lives, which helps me to relate to them a little bit more, even if they are involved in crime. Besides, I find it great that the series portrays these virtues of partnership, fidelity, and selflessness, thus turning it into an emotional experience instead of just a heist story.